Organize new communicator files
