+++ app/src/main/java/waiter/Communicator/Reporter.java
@@ -1,4 +1,4 @@
-package waiter;
+package waiter.Communicator;
 
 import java.io.IOException;
