+++ app/src/main/java/waiter/EchoServer.java
@@ -1,5 +1,7 @@
 package waiter;
 
+import waiter.Communicator.Reporter;
+
 import java.io.IOException;
 import java.net.SocketException;
